+++ warehouse/types.go
@@ -1,8 +1,54 @@
 package warehouse
 
 import (
+	"encoding/json"
+	"sync"
 	"time"
+
+	"github.com/rudderlabs/rudder-server/utils/misc"
+	warehouseutils "github.com/rudderlabs/rudder-server/warehouse/utils"
 )
+
+type PayloadT struct {
+	BatchID             string
+	UploadID            int64
+	StagingFileID       int64
+	StagingFileLocation string
+	Schema              map[string]map[string]string
+	SourceID            string
+	DestinationID       string
+	DestinationType     string
+	DestinationConfig   interface{}
+	LoadFileIDs         []int64
+}
+
+type ProcessStagingFilesJobT struct {
+	Upload    warehouseutils.UploadT
+	List      []*StagingFileT
+	Warehouse warehouseutils.WarehouseT
+}
+
+type LoadFileJobT struct {
+	Upload                     warehouseutils.UploadT
+	StagingFile                *StagingFileT
+	Schema                     map[string]map[string]string
+	Warehouse                  warehouseutils.WarehouseT
+	Wg                         *misc.WaitGroup
+	LoadFileIDsChan            chan []int64
+	TableToBucketFolderMap     map[string]string
+	TableToBucketFolderMapLock *sync.RWMutex
+}
+
+type StagingFileT struct {
+	ID           int64
+	Location     string
+	SourceID     string
+	Schema       json.RawMessage
+	Status       string // enum
+	CreatedAt    time.Time
+	FirstEventAt time.Time
+	LastEventAt  time.Time
+}
 
 type BatchRouterEventT struct {
 	Metadata MetadataT `json:"metadata"`
